Remove address finder service
Having review the code that uses the address finder service, it is commonly checking for an array, which it interprets as a success, or then for a symbol to tell it what has gone wrong.

In essence it needs the response object the `defra_ruby_address` gem returns so it can perform these checks and determine what action to take next.

Having reviewed the address finder service, its primary purpose was to make the request to the address lookup service, and then formulate its own `response` based on the result. So it now seems pointless hence this change removes it completely.

Subsequent changes will update the existing code to just call the lookup in the gem directly.