Resolve issue with failing address tests
The application works (manual testing), but because of a slight tweak to how the request is now made VCR was seeing it as a different request. This was causing the test suite to fail.

Manually tweaked the cassettes and now they match, VCR recognises them and should update them during the next VCR update.