Use defra_ruby_address for address lookup (#405)
One thing all our ruby digital services share is the need to look up addresses based on a postcode. It's a need for users to enter an address as quickly and simply as possible. And it's a need for the organisation to capture as accurate an address as possible.

What this has led to though is a lot of duplication. So we have created the [Defra Ruby Address](https://github.com/DEFRA/defra-ruby-address) gem as a way of centralising and maintaining the code which performs the lookup in one place.

This change covers removing the existing lookup code and switching to using the gem instead.