Tweak email route mounting
After crashing round in the dark trying to get this to work spotted my obvious mistake.

defra-ruby-email will provide the route `/last-email`. We need to provide what path to mount that route onto. So prior to this change the correct way to access the functionality would have been `/last-email/last-email`. Blurgh!

So correcting my mistake and hoping for better results when the WEX engine is then mounted into an app.