Issue with code coverage for SonarQube
In https://github.com/DEFRA/waste-exemptions-engine/pull/289 we implemented the ability to run the tests in parallel to reduce the build time. This worked under CodeClimate and it was able to still pick up the test coverage accurately.

However that's not the case for SonarCloud. Its is misrepesenting the code coverage as 83% when it shoul be 98%. A test done without using parallel tests resulted in accurate test coverage being calculated by SonarCloud.

We think its something about how the results.json is generated by CodeClimate when parallel tests are running, but until we can figure it out, we'll have to revert to how travis setup and ran the tests before we used parallel.