Switch to SonarCloud from CodeClimate (#441)
https://sonarcloud.io/organizations/defra/projects

We currently use [Code Climate](https://codeclimate.com/) for our code quality checks and test coverage analysis. However a number of Defra projects have used an internal SonarQube instance, and some public ones have been using SonarCloud (the cloud SASS version of SonarQube).

It looks like we're about to formally adopt SonarCloud and all projects will coalesce onto. So to keep on top of things and continue to be an example to others, we are working through our repos and switching them to SonarCloud.

** Notes

The figuring out of how to do this (because it wasn't straight forward!) was done in

- https://github.com/DEFRA/waste-exemptions-front-office/pull/317
- https://github.com/DEFRA/waste-exemptions-front-office/pull/318

* Issue with code coverage for SonarQube

In https://github.com/DEFRA/waste-exemptions-engine/pull/289 we implemented the ability to run the tests in parallel to reduce the build time. This worked under CodeClimate and it was able to still pick up the test coverage accurately.

However, that's not the case for SonarCloud. Its is misrepresenting the code coverage as 83% when it should be 98%. A test without using parallel tests resulted in accurate test coverage being calculated by SonarCloud.

We think its something about how the results.json is generated by CodeClimate when parallel tests are running, but until we can figure it out, we'll have to revert to how Travis set up and ran the tests before we used parallel.