Resolve error caused by rubocop fix
When we made the change from `.keys.each` to `.each_key` one test passed and one started erroring.

```text
    RuntimeError:
       can't add a new key into hash during iteration
     # ./app/models/waste_exemptions_engine/transient_address.rb:36:in `create_address'
     # ./app/models/waste_exemptions_engine/transient_address.rb:32:in `create_from_grid_reference_data'
     # ./spec/models/waste_exemptions_engine/transient_address_spec.rb:104:in `block (3 levels) in <module:WasteExemptionsEngine>'
     # ./spec/models/waste_exemptions_engine/transient_address_spec.rb:108:in `block (4 levels) in <module:WasteExemptionsEngine>'
     # ./spec/models/waste_exemptions_engine/transient_address_spec.rb:107:in `each_key'
     # ./spec/models/waste_exemptions_engine/transient_address_spec.rb:107:in `block (3 levels) in <module:WasteExemptionsEngine>'
```

After some investigation, it was found that it's because

> Ruby since 1.9 is using lazy iteration, thus they forbid to add new keys to the hash you iterating over.
Ref: https://stackoverflow.com/a/27643890/6117745

In the working test we had to hashs. One was passed to the method under test, and the other was iterated over in the spec. In the non-working test the same has was passed to the method and iterated over. The method under test amends the hash passed in hence we get the error.

We could have reverted the change and added a rubocop rule exemption. However I've chosen to update the failing test to follow the pattern used in the working one. It felt to me this made things more consistent, and in a way clearer as to what is actually going on and what we are testing for.