Don't expire VCR cassettes
We still want to keep VCR in this project. However, re-recording every fortnight is too big a maintenance overhead,   and we have other ways of detecting breaking changes now.

This change removes the automatic two-week expiry, so cassettes will remain as they are until the request is modified
or we decide to update them.