Add ruby-prof to gemspec
This is an helpful tool usually used to debug issues with the ruby code.
In our case it has been used to debug the possible memory leak issue reported on https://github.com/thoughtbot/factory_bot_rails/issues/384 hence we might want to keep the code around for future reference.