Updgrade Feature Toggle to proper model
From: https://eaflood.atlassian.net/browse/RUBY-1139

This upgrade the feature toggle to be a proper model.
It also fixes the content of the `bin/rails` file which was pointing to the wrong files.
The old Feature Toggle functionality and interface as stayied the same. To achieve 0 downtime we will fall back to the content of the yml if we can't find a match for a given feature on the database. Once this is merged and deployed, we can create the records using the back-office interface and then at the subsequent deploy we can cleanup the code in the model that fallsback to the content of the config file.