Updgrade Feature Toggle to proper model (#480)
https://eaflood.atlassian.net/browse/RUBY-1139

Given that little `FeatureToggle` has behaved very well in the past few months, we decided to upgrade it to be a proper Model 😁 

This model will be used by the new [defra-ruby-features](https://github.com/DEFRA/defra-ruby-features) engine. Toggles will be persisted in the database.

Things to notice are, in order to handle the fact that the model won't have persisted data about existing toggles which were previously persisted in a yml file, the code will handle that case by redirecting the query for a toggle status to the content of the yml file.

After we deploy we'll just need to take our time creating the toggles through the back-office interface. In the meantime, everything will work just fine.

Once deployed and the existing features are migrated from the yml file to the database we can begin to clean up the code. But we still think this is a cleaner solution than a one-time rake task and Jenkins job.

N.B. Also fixes the content of the `bin/rails` file which was pointing to the wrong files.

Co-authored-by: Alan Cruikshanks <alan.cruikshanks@gmail.com>