Set up ConfirmationEmailService
https://eaflood.atlassian.net/browse/RUBY-1183

This is the service that will replace the ConfirmationMailer. It sends a request to Notify to send out the new confirmation email.

Previously we had a fallback in case the certificate PDF wasn't generated properly. It's not possible to handle this within the same Notify template so instead we request a different email template depending on whether the PDF is generated successfully or not.