Fix confirmation email error handling
https://eaflood.atlassian.net/browse/RUBY-1186

We ran into a bug where if the applicant email wasn't on the guest list, the contact email wouldn't be emailed either.

This turned out to be because we rescued the error from the applicant email in the wrong place. This update fixes it so we will still try the second email if the first fails.