Move Notify confirmation letter code from BO to engine (#536)
https://eaflood.atlassian.net/browse/RUBY-1248

In order to send letters when registrations are completed, rather than on an overnight schedule, we need to be able to send them from *wherever* they are completed. This means both front and back office, so it's time to move this code into the engine.

This had a lot of dependencies on other bits of code so I had to move that over too, along with some slight refactoring to avoid deprecation warnings.