Configure AD email in engine initialiser (#538)
https://eaflood.atlassian.net/browse/RUBY-1208

Previously the AD email address was hardcoded into the back office.

This is a pain for testing and flexibility, and inconsistent with waste carriers, so we should allow it to be configured in the engine initialiser instead. This will let us set a default if we like or override it with an environment variable.