Adding 'Error' in the page title when an error is generated  (#639)
https://eaflood.atlassian.net/jira/software/projects/RUBY/boards/374?selectedIssue=RUBY-1727

Here we have updated the page titles to change when an error is generated. This allows screen readers to give a better description to the user on the pages status.

The titles will now read <% Error - What do you want to do? - Register your waste exemptions - GOV.UK %> when an error has been generated.
Four test has been written in application_helper_spec to check the helper method for titles and to check the errors in titles.