Playback operator address to use as site address (#661)
https://eaflood.atlassian.net/browse/RUBY-1843

Here we are making changes so the user doesn't have to manually re-enter their site address if they have already entered it previously as either their business address or contact address. If the address is different from these ones then they have the opportunity to enter a new one or the grid reference.