Companies House data retrieval  (#665)
https://eaflood.atlassian.net/browse/RUBY-1872

Here we have the users company house number being used to validate that the company is active and return the company name and address for the user to confirm it is indeed their company. If the user confirms this then the name name registered on company's house will be used as the operator name for the service. If the user doesn't confirm these details then they are asked to change them via company house