New registration relabel operator name on check answers page (#672)
https://eaflood.atlassian.net/browse/RUBY-1873
This is a label change for the check your answers page. When the business type is either a LLP or company the label is changed to Registered Name (as this is pulled from companies house and isn't entered by the user) however if its anything else i.e. charity then the label is Operator Name.

Spec file has been changed to test the new method and to test a charity on the 'check your answers page' as well.